refactor: step3 패키지 분리
