+++ src/main/java/bowling/domain/score/Score.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 import java.util.Objects;
