+++ src/test/java/bowling/domain/NameTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.player.Name;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
