+++ src/test/java/bowling/domain/PinTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.pin.Pin;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
