+++ src/test/java/bowling/domain/PinsTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.pin.Pins;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
