+++ src/test/java/bowling/domain/PitchTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.pitch.Pitch;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
