+++ src/test/java/bowling/domain/PlayerTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.player.Player;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
